bigquery_sink_agent/requirements.txt
google-cloud-bigquery
fastapi
uvicorn[standard]